+++ src/components/settings/TimerSettings.tsx
@@ -12,6 +12,7 @@ import { AmbientSettings } from './AmbientSettings';
 import { VisualEffectsSettings } from './VisualEffectsSettings';
 import { CustomPresetEditor } from './CustomPresetEditor';
 import { WeekStartSetting } from './WeekStartSetting';
+import { DataExportSettings } from './DataExportSettings';
 import { OverflowSettings } from './OverflowSettings';
 import { AutoStartSettings } from './AutoStartSettings';
 import { EndTimeSettings } from './EndTimeSettings';
@@ -312,6 +313,11 @@ export function TimerSettings({ onSettingsChange, disabled }: TimerSettingsProps
 
                   {/* Week Start Setting */}
                   <WeekStartSetting />
+
+                  {/* ═══ Privacy Section ═══ */}
+                  <div className="pt-4 mt-4 border-t border-tertiary/10 light:border-tertiary-dark/10">
+                    <DataExportSettings />
+                  </div>
                 </div>
 
                 {/* Footer hint */}